+++ Woof/Woof/View/SitterFlow/SitterWalkingsView/SitterWalkingsView.swift
@@ -4,11 +4,56 @@ import SwiftUI
 struct SitterWalkingsView: View {
     var body: some View {
         NavigationView {
-            Text("Walkings stub")
-                .navigationTitle("Walkings")
-                .navigationBarTitleDisplayMode(.inline)
+            ScrollView {
+                SitterWalkingsSectionView(
+                    isExpanded: showCurrentWalkings,
+                    walkings: viewModel.currentWalkings,
+                    headerTitle: currentWalkingSectionTitle
+                )
+
+                SitterWalkingsSectionView(
+                    isExpanded: showAllFutureWalkings,
+                    walkings: viewModel.futureWalkings,
+                    headerTitle: futureWalkingSectionTitle
+                )
+
+                SitterWalkingsSectionView(
+                    isExpanded: showAllFinishedWalkings,
+                    walkings: viewModel.finishedWalkings,
+                    headerTitle: finishedWalkingSectionTitle
+                )
+
+                Spacer()
+            }
+            .overlay {
+                if viewModel.isWalkingsLoading {
+                    CustomProgressView()
+                }
+            }
+            .padding()
+            .listStyle(.plain)
+            .navigationTitle(navigationTitle)
+            .navigationBarTitleDisplayMode(.inline)
+            .refreshable {
+                Task {
+                    await viewModel.getWalkings()
+                }
+            }
         }
     }
+
+    // MARK: - Private interface
+
+    @State private var showAllFutureWalkings = true
+    @State private var showCurrentWalkings = true
+    @State private var showAllFinishedWalkings = false
+
+    @StateObject private var viewModel = SitterWalkingsViewModel()
+
+    private let futureWalkingSectionTitle = "Future walkings"
+    private let currentWalkingSectionTitle = "Current walkings"
+    private let finishedWalkingSectionTitle = "Finished walkings"
+    private let navigationTitle = "Walkings"
 }
 
 struct SitterWalkingsView_Previews: PreviewProvider {